Moved almost all Entity logic to a separate GameLoop class
