+++ core/src/main/java/core/game/logic/GameLogic.java
@@ -1,0 +1,45 @@
+package core.game.logic;
+
+import com.badlogic.gdx.Gdx;
+
+import java.util.Timer;
+import java.util.TimerTask;
+
+public class GameLogic {
+
+    private static Timer gameTimer;
+
+    public static void start() {
+
+        gameTimer = new Timer();
+        gameTimer.schedule(new TimerTask() {
+            @Override
+            public void run() {
+                gameTick();
+            }
+        }, Entity.TIC);
+    }
+
+    public static void stop() {
+
+        gameTimer.cancel();
+        gameTimer.purge();
+    }
+
+    private static void gameTick() {
+        for (Entity e : Entity.entityList) {
+            e.decrementTics();
+
+            if (e instanceof PlayerPawn) {
+                ((PlayerPawn) e).movementUpdate();
+            }
+        }
+
+        gameTimer.schedule(new TimerTask() {
+            @Override
+            public void run() {
+                gameTick();
+            }
+        }, Entity.TIC);
+    }
+}
