+++ core/src/main/java/core/gdx/wad/MyGDxTest.java
@@ -3,6 +3,9 @@ package core.gdx.wad;
 import com.badlogic.gdx.Game;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.audio.Sound;
+import core.game.logic.Entity;
+import core.game.logic.GameLogic;
+import core.game.logic.PlayerPawn;
 
 import static com.badlogic.gdx.Gdx.audio;
 
@@ -13,8 +16,7 @@ public class MyGDxTest extends Game {
     @Override
     public void create() {
 
-
-        gameScreen = new GameScreen(this);
+        gameScreen = new GameScreen();
         setScreen(gameScreen);
     }
 }
